+++ src/com/intellij/idea/plugin/hybris/system/cockpitng/psi/reference/CngWidgetReference.kt
@@ -1,6 +1,6 @@
 /*
- * This file is part of "SAP Commerce Developers Toolset" plugin for Intellij IDEA.
- * Copyright (C) 2019 EPAM Systems <hybrisideaplugin@epam.com>
+ * This file is part of "SAP Commerce Developers Toolset" plugin for IntelliJ IDEA.
+ * Copyright (C) 2019-2024 EPAM Systems <hybrisideaplugin@epam.com> and contributors
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU Lesser General Public License as
@@ -22,20 +22,38 @@ import com.intellij.codeInsight.highlighting.HighlightedReference
 import com.intellij.idea.plugin.hybris.psi.util.PsiUtils
 import com.intellij.idea.plugin.hybris.system.cockpitng.meta.CngMetaModelAccess
 import com.intellij.idea.plugin.hybris.system.cockpitng.psi.reference.result.WidgetResolveResult
+import com.intellij.openapi.util.Key
 import com.intellij.psi.PsiElement
 import com.intellij.psi.PsiPolyVariantReference
 import com.intellij.psi.PsiReferenceBase
 import com.intellij.psi.ResolveResult
+import com.intellij.psi.util.*
 
 class CngWidgetReference(element: PsiElement) : PsiReferenceBase.Poly<PsiElement>(element), PsiPolyVariantReference, HighlightedReference {
 
-    override fun multiResolve(incompleteCode: Boolean): Array<ResolveResult> {
-        val lookingForName = value
+    override fun multiResolve(incompleteCode: Boolean): Array<ResolveResult> = CachedValuesManager.getManager(element.project)
+        .getParameterizedCachedValue(element, CACHE_KEY, provider, false, this)
+        .let { PsiUtils.getValidResults(it) }
 
-        return CngMetaModelAccess.getInstance(element.project).getMetaModel()
-            .widgets[lookingForName]
-            ?.let { PsiUtils.getValidResults(arrayOf(WidgetResolveResult(it))) }
-            ?: emptyArray()
+    companion object {
+        val CACHE_KEY = Key.create<ParameterizedCachedValue<Array<ResolveResult>, CngWidgetReference>>("HYBRIS_CNGWIDGETREFERENCE")
+
+        private val provider = ParameterizedCachedValueProvider<Array<ResolveResult>, CngWidgetReference> { ref ->
+            val element = ref.element
+            val lookingForName = ref.value
+            val project = element.project
+            val metaModel = CngMetaModelAccess.getInstance(project).getMetaModel()
+
+            val result = metaModel
+                .widgets[lookingForName]
+                ?.let { PsiUtils.getValidResults(arrayOf(WidgetResolveResult(it))) }
+                ?: emptyArray()
+
+            CachedValueProvider.Result.create(
+                result,
+                metaModel, PsiModificationTracker.MODIFICATION_COUNT
+            )
+        }
     }
 
 }